push-commits: add support for RelBranches (Bug 1964716)
Add a `--relbranch` flag to `push-commits` which allows creating new
RelBranches on the designated repo. When passed, `--relbranch` creates
a new RelBranch when one could not be detected, or adds commits to
an existing detected relbranch. When a new RelBranch is created, a
warning is displayed about the new branch not yet tracking the
upstream, with instructions on how to set the local branch to track
the remote.