add-commits: use bytes when dealing with format-patch output (bug 1967655) (#11)
This allows us to handle DOS line terminators. Otherwise, any line
ending in patch output is transformed into Unix-style terminators, which
later leads to conflicts when trying to apply the patch. Gather patches as
raw `bytes`, then encode as `base64` and switch to using the `add-commit-base64`
action to submit patches.

---------

Co-authored-by: Connor Sheehan <cosheehan@mozilla.com>